requirements.txt
pythodiscord.py
sgfmill
cairosvg
APScheduler
